test-Clip-Clap/requirements.txt
// ho usato python 3.10

accelerate==0.21.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
bitsandbytes==0.45.5
Bottleneck==1.4.2
braceexpand==0.1.7
Brotli==1.0.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cramjam==2.10.0
cycler==0.11.0
debugpy==1.8.11
decorator==5.2.1
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.115.12
fastparquet==2024.11.0
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.55.3
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
h5py==3.13.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.31.2
idna==3.7
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
laion_clap==1.1.6
latex2mathml==3.78.0
lazy_loader==0.4
librosa==0.10.2.post1
llava==1.2.2.post1
llvmlite==0.41.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
msclap==1.3.3
msgpack==1.0.3
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.58.1
numexpr==2.10.1
numpy==1.26.4
openai-clip==1.0.1
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pillow==10.4.0
pip==25.1
platformdirs==4.3.8
pooch==1.8.2
progressbar==2.5
prompt-toolkit==3.0.43
protobuf==6.31.0
psutil==5.9.0
pure-eval==0.2.2
pyarrow==20.0.0
pycparser==2.21
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==14.0.0
rpds-py==0.25.1
ruff==0.11.11
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==72.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spyder-kernels==2.5.0
stack-data==0.2.0
standard-aifc==3.13.0
standard-sunau==3.13.0
starlette==0.46.2
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchlibrosa==0.1.0
torchvision==0.16.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
typer==0.15.4
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
unicodedata2==15.1.0
urllib3==2.3.0
uvicorn==0.34.2
wandb==0.19.11
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.111
websockets==11.0.3
wget==3.2
wheel==0.45.1
win-inet-pton==1.1.0